--- backend/src/main/java/archloner/todo/ApplicationConfig.java
@@ -1,8 +1,0 @@
-package archloner.todo;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ApplicationConfig {
-
-}
