--- backend/src/main/java/archloner/todo/data/ProjectRepository.java
@@ -1,7 +1,0 @@
-package archloner.todo.data;
-
-import archloner.todo.model.Project;
-import org.springframework.data.repository.CrudRepository;
-
-public interface ProjectRepository extends CrudRepository<Project, Long> {
-}
