--- backend/src/main/java/archloner/todo/data/TaskRepository.java
@@ -1,7 +1,0 @@
-package archloner.todo.data;
-
-import archloner.todo.model.Task;
-import org.springframework.data.repository.CrudRepository;
-
-public interface TaskRepository extends CrudRepository<Task, Long> {
-}
