--- backend/src/main/java/archloner/todo/data/UserRepository.java
@@ -1,10 +1,0 @@
-package archloner.todo.data;
-
-import archloner.todo.model.User;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-
-    User findByUsername(String username);
-
-}
