--- backend/src/main/java/archloner/todo/model/Project.java
@@ -1,30 +1,0 @@
-package archloner.todo.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-
-import java.util.Date;
-import java.util.List;
-
-@Data
-@RequiredArgsConstructor
-@Entity
-public class Project {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long projectId;
-    private String name;
-    private String description;
-
-    @OneToMany
-    private List<Task> taskList;
-
-    @CreatedDate
-    private Date createdAt;
-
-    private User ownerUser;
-
-}
