--- backend/src/main/java/archloner/todo/model/TaskPriority.java
@@ -1,7 +1,0 @@
-package archloner.todo.model;
-
-public enum TaskPriority {
-
-    DEFAULT_PRIORITY, LOW_PRIORITY, MEDIUM_PRIORITY, HIGHT_PRIORITY
-
-}
