--- backend/src/main/java/archloner/todo/security/TodoUserDetailsService.java
@@ -1,27 +1,0 @@
-package archloner.todo.security;
-
-import archloner.todo.data.UserRepository;
-import archloner.todo.model.User;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-public class TodoUserDetailsService implements UserDetailsService {
-
-    private UserRepository userRepo;
-
-    @Autowired
-    public TodoUserDetailsService(UserRepository userRepo) {
-        this.userRepo = userRepo;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepo.findByUsername(username);
-        if (user != null) {
-            return user;
-        }
-        throw new UsernameNotFoundException("User {username} not found");
-    }
-}
