--- backend/src/main/java/archloner/todo/web/api/TaskController.java
@@ -1,31 +1,0 @@
-package archloner.todo.web.api;
-
-import archloner.todo.data.TaskRepository;
-import archloner.todo.model.Task;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
-
-@RestController
-@RequestMapping(path = "/api/task", produces = "application/json")
-@CrossOrigin(origins = { "localhost:8080" })
-public class TaskController {
-
-    private TaskRepository taskRepository;
-
-    @Autowired
-    public TaskController(TaskRepository taskRepository) {
-        this.taskRepository = taskRepository;
-    }
-
-    @GetMapping("/{taskId}")
-    public ResponseEntity<Task> getTaskById(@PathVariable("taskId") Long taskId) {
-        Optional<Task> taskOpt = taskRepository.findById(taskId);
-        return taskOpt.map(task -> new ResponseEntity<>(task, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(null, HttpStatus.NOT_FOUND));
-    }
-
-}
