--- backend/src/main/resources/application.yml
@@ -1,61 +1,0 @@
-spring:
-  profiles:
-    active: dev-mac
-  application:
-    name: TodoApplication
-
----
-spring:
-  config:
-    activate:
-      on-profile: dev-mac
-  datasource:
-    url: jdbc:mysql://localhost:3306/todo_db
-    username: todoapp
-    password: password
-    driverClassName: com.mysql.cj.jdbc.Driver
-    initialization-mode: always
-  jpa:
-    show-sql: true
-    hibernate:
-      ddl-auto: update
-
-devtools:
-  livereload:
-    enabled: true
-
-server:
-  port: 8088
-
-logging:
-  level:
-    org.hibernate.SQL: debug
-    org.hibernate.type.descriptor.sql.BasicBinder: trace
-
-
----
-spring:
-  config:
-    activate:
-      on-profile: dev-win
-  datasource:
-    url: jdbc:mysql://localhost:3306/todo_db
-    username: todoapp
-    password: password
-    driverClassName: com.mysql.cj.jdbc.Driver
-  jpa:
-    show-sql: true
-
-
-devtools:
-  livereload:
-    enabled: true
-
-server:
-  port: 8088
-
----
-spring:
-  config:
-    activate:
-      on-profile: prod
