--- backend/src/test/java/archloner/todo/TodoApplicationTests.java
@@ -1,13 +1,0 @@
-package archloner.todo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TodoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
