+++ backend/todoism-api/src/main/java/tdsm/TaskController.java
@@ -1,0 +1,28 @@
+package tdsm;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import tdsm.domain.Task;
+
+@RestController
+@RequestMapping(path = "/api", produces = "application/json")
+public class TaskController {
+
+    @Slf4j
+    private static class TaskControllerLogger {}
+
+    public TaskController(TaskRepository taskRepo) {
+        this.taskRepo = taskRepo;
+    }
+
+    public TaskRepository taskRepo;
+
+    @GetMapping("/task")
+    public Iterable<Task> findAllTasks() {
+        TaskControllerLogger.log.info("Returning list of all tasks in TaskController");
+        return taskRepo.findAll();
+    }
+
+}
