+++ backend/todoism-data/src/main/java/tdsm/ProjectRepository.java
@@ -1,0 +1,7 @@
+package tdsm;
+
+import org.springframework.data.repository.CrudRepository;
+import tdsm.domain.Project;
+
+public interface ProjectRepository extends CrudRepository<Project, Long> {
+}
