+++ backend/todoism-data/src/main/java/tdsm/TaskRepository.java
@@ -1,0 +1,7 @@
+package tdsm;
+
+import org.springframework.data.repository.CrudRepository;
+import tdsm.domain.Task;
+
+public interface TaskRepository extends CrudRepository<Task, Long> {
+}
