+++ backend/todoism-data/src/main/java/tdsm/UserRepository.java
@@ -1,0 +1,9 @@
+package tdsm;
+
+import org.springframework.data.repository.CrudRepository;
+
+//public interface UserRepository extends CrudRepository<User, Long> {
+//
+//    User findByUsername(String username);
+//
+//}
