+++ backend/todoism-domain/src/main/java/tdsm/domain/TaskPriority.java
@@ -1,0 +1,7 @@
+package tdsm.domain;
+
+public enum TaskPriority {
+
+    DEFAULT_PRIORITY, LOW_PRIORITY, MEDIUM_PRIORITY, HIGHT_PRIORITY
+
+}
