+++ backend/todoism/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+logging.level.org.springframework.data.mongodb.core=DEBUG
+spring.output.ansi.enabled=ALWAYS
+org.springframework.boot.devtools.classpath=DEBUG
