+++ backend/todoism/src/main/resources/application.yml
@@ -1,0 +1,8 @@
+spring:
+  data:
+    mongodb:
+      uri: mongodb://127.0.0.1:27017/todolistDB
+
+devtools:
+  livereload:
+    enabled: true
